backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
requests==2.32.3
pytest==8.3.2
pytest-django==4.8.0
httpx==0.27.2
